Added index template, tidied up classes, fixed tests
Changing CurrencyExchangeController to use @Controller annotation
caused mappings that returned a String to fail. For @Controller
to work there has to be a corresponding html template.

Also tidied up some of the classes, removing unnecessary
constructors and lines.

Also fixed CurrencyExchangeControllerTest; this was checking for
for the object to be NotNull - this is wrong because even with
a failure, the fail response is returned as a String so is never
null. Changed test to check the returned html/String does not
contain "error" or "404".